requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
Pillow>=10.3.0
pix2tex[api]>=0.1.2
pydantic>=2.7.1
httpx>=0.27.0
pytest>=8.2.0
pytest-asyncio>=0.23.6
